+++ src/main/java/com/modeunsa/boundedcontext/payment/app/usecase/PaymentCreditAccountUseCase.java
@@ -2,6 +2,7 @@ package com.modeunsa.boundedcontext.payment.app.usecase;
 
 import com.modeunsa.boundedcontext.payment.app.support.PaymentAccountSupport;
 import com.modeunsa.boundedcontext.payment.domain.types.PaymentEventType;
+import java.math.BigDecimal;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,7 +18,7 @@ public class PaymentCreditAccountUseCase {
   private final PaymentAccountSupport paymentAccountSupport;
 
   @Transactional
-  public void execute(Long memberId, long amount, PaymentEventType paymentEventType) {
+  public void execute(Long memberId, BigDecimal amount, PaymentEventType paymentEventType) {
     var paymentAccount = paymentAccountSupport.getPaymentAccountByMemberId(memberId);
     paymentAccount.credit(amount, paymentEventType);
   }
